+++ src/types/requests/queries.ts
@@ -7,6 +7,8 @@ export type FilterValue = string | number | boolean | string[] | number[] | bool
 export type PositiveFilter = { [field: string]: FilterValue };
 export type NegativeFilter = { $not?: PositiveFilter };
 export type Filter = PositiveFilter | NegativeFilter;
+//TODO: BqlFilter, EnrichedBQLFilter, then in the adapters, the TqlParsedFilter and BqlParsedFilter
+//TODO: Refacto of Filter types for more flexibility, $or, $and etc
 
 export type RawBQLQuery = {
 	$id?: string | string[];